+++ Components/DownloadMenager/WebWrapper.js
@@ -1,0 +1,5 @@
+function drawMod() {
+  console.log(this.Dobjects);
+}
+
+module.exports = drawMod;
